+++ R/init.R
@@ -109,6 +109,9 @@ init <- function(data,
   # ui function
   ui <- shinyUI(
       fluidPage(
+        shinyjs::useShinyjs(),
+        includeScript(system.file("js/clipboard.js", package = "teal")),
+        includeScript(system.file("js/initClipboard.js", package = "teal")),
         tags$head(
           tags$script(
             # show/hide see https://groups.google.com/forum/#!topic/shiny-discuss/yxFuGgDOIuM